+++ routes/console_routes/conduit_routes.php
@@ -72,16 +72,3 @@ Artisan::command('conduit:check-discord-error', function () {
 })->purpose('Sends an notification to test the Conduit Discord Error Logger');
 
 
-/**
- * Test Web Request
- */
-Artisan::command('conduit:check-web-request', function () {
-  $this->comment('Checking the Conduit Web Request');
-
-  $qbt_instance = new QuickbooksTime();
-  $result = $qbt_instance->getQBTUsers();
-
-  $this->comment($result);
-
-})->purpose('Checks that web requests are working within Conduit');
-