+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This project provides a central API and Dashboard for communication between appl
 - Adds Users Sync from Dolibarr (FirstName, Lastname, Email, Password, Username, JSON Blob of Object)
 - Adds Projects Sync from Dolibarr
 - Adds Workorder model and  Sync from Dolibarr to Conduit
+- Adds User Syncing with Quickbooks Time
+- Adds project syncing with Quickbooks Time
 ### Changes
 ### Removes
 ### Fixes